api/requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
gunicorn==19.9.0
idna==2.8
imutils==0.5.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
numpy==1.16.2
opencv-python==4.0.0.21
pymongo==3.7.2
python-dateutil==2.8.0
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.15.2
